+++ tests/utils.py
@@ -14,8 +14,18 @@ from os.path import join
 DIR_CASES = join('.', 'tests', 'resources', 'cases')
 DIR_PFSOL = join('.', 'tests', 'resources', 'pf_solutions')
 
-test_cases = [join(DIR_CASES, f) for f in listdir(DIR_CASES)]
+SPECIAL_CASES = ['ieee25_mvar_in.raw', 'ieee25_mvar_out_limit.raw', 'ieee25_mvar_out.raw', 'ieee25_mw.raw', 'ieee25_wind.raw']
+
+test_cases = [join(DIR_CASES, f) for f in listdir(DIR_CASES) if f not in SPECIAL_CASES]
 test_cases.sort()
+
+skip_cases = ['case6_transfer.raw']
+
+# Remove skip cases from lists
+for case in skip_cases:
+    case_with_path = os.path.join(DIR_CASES, case)
+    if case_with_path in test_cases:
+        test_cases.remove(case_with_path)
 
 def get_pf_solution_file(case_file, dir, sol):
     """