+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-Verions 1.4.1
+Version 1.5.0
+-------------
+* Added variables on Inter-Area Transfer.
+* Added constraints and heuristics on regulating inter-area transfer MW.
+* Added transformer P, Q control constraints and heuristics in NR.
+* Added transformer P, Q control functions in Opt-based ACPF.
+* Added phase shift variables for phase shifters.
+* Added asymmetric transformer equation on ratio and phase angle in Opt-based.
+* Added support for fixed power factor generator controls (e.g., wind machines).
+* Added support for Q limit regulation from generator power factors (e.g., wind machines).
+* Add FACTS constraints with for series voltage control and series impedance control.
+* Added `loads_2_ZIP` parameter for ACPF to convert loads to constant impedance if voltage drops below threshold (default is False).
+* Adding admittance correction parameter `y_correction` and callback to ACPF Newton-Raphson only.
+* Added tap-changer rounding parameter `taps_round` that will round taps each step if tap-mode is regulating.
+* Moved targeted support to Python 3.10
+* Move from nosetest to pytest
+* Add tox support for py37, py38, py39, py310
+* Added KLU for DCPF
+* Changed the switched shunt ACPF from the band regulation using PFNET to a callback function that mimics PSSE that includes switched shunt operation limit check to control for oscillations.
+
+Version 1.4.1
 -------------
 * Updated setup.py to use compiled files with `bdist_wheel_compiled` and `bdist_egg_compiled`.
 * Added `load_q_curtail` parameter setting to ACPF that allows for load.Q to deviate.